requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
schedule==1.2.1
google-generativeai==0.3.2
supabase==2.0.3 